autodeletion for old socketmessages(2 week) + chatroom(if empty) added

+++ src/main/java/com/cobong/yuja/config/AutomaticFileDeletion.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Component;
 
 import com.cobong.yuja.service.board.BoardAttachService;
 import com.cobong.yuja.service.board.ThumbnailService;
+import com.cobong.yuja.service.chat.ChatRoomService;
 import com.cobong.yuja.service.chat.SocketMessageService;
 import com.cobong.yuja.service.user.ProfilePictureService;
 import com.cobong.yuja.service.user.UserService;
@@ -24,15 +25,18 @@ public class AutomaticFileDeletion {
 	private final UserService userService;
 	private final YoutubeConfirmService youtuberConfirmService;
 	private final SocketMessageService socketMessageService;
+	private final ChatRoomService chatRoomService;
 	
-	@Scheduled(cron = "0 25 8 * * *")
+	@Scheduled(cron = "0 0 4 * * *")
 	public void deleteAtFourAM() {
 		File tempToDel = Paths.get(System.getProperty("user.dir") + File.separator+"files" + File.separator +"temp").toFile();
 		File[] tempsToDel = tempToDel.listFiles();
-		for(File file: tempsToDel) {
-			file.delete();			
+		if(tempsToDel != null && tempsToDel.length != 0) {
+			for(File file: tempsToDel) {
+				file.delete();			
+			}
+			tempToDel.delete();
 		}
-		tempToDel.delete();
 		
 		boardAttachService.deleteUnflagged();
 		profilePictureService.deleteUnflagged();
@@ -40,5 +44,6 @@ public class AutomaticFileDeletion {
 		userService.deleteUnflagged();
 		youtuberConfirmService.deleteUnflagged();
 		socketMessageService.delete2weeksOld();
+		chatRoomService.deleteEmptyRooms();
 	}
 }
